app-store-screen-shot/.claude/skills/appstore-screenshot/requirements.txt
google-genai>=0.8.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
